+++ meal-planner/src/components/RecipePage.md
@@ -1,0 +1,6 @@
+<RecipePage
+title="Ein Gericht"
+image="../public/images/risotto.png"
+ingredients="Zutaten"
+instructions="Text"
+/>
