+++ meal-planner/src/data/testShoppingItems.js
@@ -1,0 +1,12 @@
+const shoppingItems = [
+    { id: '1', ingredient: '1 Dose Tomaten' },
+    { id: '2', ingredient: '1 Stange Lauch' },
+    { id: '3', ingredient: '1 Packung Nudeln' },
+    { id: '4', ingredient: '1 Baguette' },
+    { id: '5', ingredient: '1 Packung Spaghetti' },
+    { id: '6', ingredient: '1 Netz Mandarinen' },
+    { id: '7', ingredient: '1 Dose Tomaten' },
+    { id: '8', ingredient: '2 Stangen Wasser' },
+]
+
+export default shoppingItems
