+++ meal-planner/styleguide.config.js
@@ -1,7 +1,12 @@
+const path = require('path')
+
 module.exports = {
     components: 'src/components/**/*.js',
-    ignore: ['**/App.js'],
+    ignore: ['**/App.js', '**/*.test.js', '**/GlobalStyle.js'],
     defaultExample: true,
     exampleMode: 'expand',
     usageMode: 'expand',
+    styleguideComponents: {
+        Wrapper: path.join(__dirname, 'src/styles/StyleWrapper'),
+    },
 }
